WIP: Reorganize structure and design
