--- resources/lib/vrtplayer/actions.py
@@ -1,17 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# GNU General Public License v3.0 (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
-
-from __future__ import absolute_import, division, unicode_literals
-
-
-LISTING_AZ_TVSHOWS = 'listingaztvshows'
-LISTING_CATEGORIES = 'listingcategories'
-LISTING_CATEGORY_TVSHOWS = 'listingcategorytvshows'
-LISTING_CHANNELS = 'listingchannels'
-LISTING_EPISODES = 'listingepisodes'
-LISTING_LIVE = 'listinglive'
-LISTING_RECENT = 'listingrecent'
-LISTING_TVGUIDE = 'listingtvguide'
-
-PLAY = 'play'
