+++ service.py
@@ -6,26 +6,25 @@
 
 from __future__ import absolute_import, division, unicode_literals
 
-import xbmc
-import xbmcaddon
-from resources.lib.kodiwrappers import kodiwrapper
-from resources.lib.vrtplayer import tokenresolver
+from vrtnu.kodiwrapper import KodiWrapper
+from vrtnu.tokenresolver import TokenResolver
+from xbmc import Monitor
+from xbmcaddon import Addon
 
 
-class VrtMonitor(xbmc.Monitor):
+class VrtMonitor(Monitor):
     ''' This is the class that monitors Kodi for the VRT NU video plugin '''
 
     def __init__(self):
         ''' VRT Monitor initialisiation '''
-        xbmc.Monitor.__init__(self)
+        Monitor.__init__(self)
 
     def onSettingsChanged(self):
         ''' Handler for changes to settings '''
-        addon = xbmcaddon.Addon(id='plugin.video.vrt.nu')
-        kodi_wrapper = kodiwrapper.KodiWrapper(None, None, addon)
+        addon = Addon(id='plugin.video.vrt.nu')
+        kodi_wrapper = KodiWrapper(None, addon)
         kodi_wrapper.log_notice('VRT NU Addon: settings changed')
-        token_resolver = tokenresolver.TokenResolver(kodi_wrapper)
-        token_resolver.reset_cookies()
+        TokenResolver(kodi_wrapper).reset_cookies()
 
 
 if __name__ == '__main__':
