+++ vrtnu/statichelper.py
