+++ vrtnu/vrtapihelper.py
@@ -3,8 +3,8 @@
 # GNU General Public License v3.0 (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, unicode_literals
-from resources.lib.helperobjects import helperobjects
-from resources.lib.vrtplayer import actions, metadatacreator, statichelper
+from vrtnu.helperobjects import actions, TitleItem
+from vrtnu import metadatacreator, statichelper
 
 try:
     from urllib.parse import urlencode, unquote
@@ -58,11 +58,13 @@ class VRTApiHelper:
             # Cut vrtbase url off since it will be added again when searching for episodes
             # (with a-z we dont have the full url)
             video_url = statichelper.add_https_method(tvshow.get('targetUrl')).replace(self._VRT_BASE, '')
-            tvshow_items.append(helperobjects.TitleItem(title=label,
-                                                        url_dict=dict(action=actions.LISTING_EPISODES, video_url=video_url),
-                                                        is_playable=False,
-                                                        art_dict=dict(thumb=thumbnail, icon='DefaultAddonVideo.png', fanart=thumbnail),
-                                                        video_dict=metadata.get_video_dict()))
+            tvshow_items.append(TitleItem(
+                title=label,
+                url_dict=dict(action=actions.LISTING_EPISODES, video_url=video_url),
+                is_playable=False,
+                art_dict=dict(thumb=thumbnail, icon='DefaultAddonVideo.png', fanart=thumbnail),
+                video_dict=metadata.get_video_dict(),
+            ))
         return tvshow_items
 
     def _get_season_items(self, api_url, api_json):
@@ -213,7 +215,7 @@ class VRTApiHelper:
             video_url = statichelper.add_https_method(episode.get('url'))
             label, sort, ascending = self._make_label(episode, titletype, options=display_options)
             metadata.title = label
-            episode_items.append(helperobjects.TitleItem(
+            episode_items.append(TitleItem(
                 title=label,
                 url_dict=dict(action=actions.PLAY, video_url=video_url, video_id=episode.get('videoId'), publication_id=episode.get('publicationId')),
                 is_playable=True,
@@ -245,7 +247,7 @@ class VRTApiHelper:
             label = '%s %s' % (self._kodi_wrapper.get_localized_string(30094), season_key)
             params = {'facets[seasonTitle]': season_key}
             path = api_url + '&' + urlencode(params)
-            season_items.append(helperobjects.TitleItem(
+            season_items.append(TitleItem(
                 title=label,
                 url_dict=dict(action=actions.LISTING_EPISODES, video_url=path),
                 is_playable=False,
